requirements.txt
# Backend Dependencies (Pip Packages)
fastapi
uvicorn

# Standard Library Modules Used (Built-in, no install needed):
# - datetime
# - time

# Frontend Technologies (Built-in, no install needed):
# - HTML5
# - CSS3 (Vanilla)
# - JavaScript (Vanilla)

# Auth Dependencies
python-jose[cryptography]
passlib[bcrypt]
multipart  

# Dev Dependencies
python-dotenv
fastapi[standard]

# Database
beanie          # Async MongoDB ODM